Move FilePrototype into fileprototype.go
